backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Authlib==1.3.0
requests==2.31.0
pandas==2.1.3
reportlab==4.0.7
python-dotenv==1.0.0
